feat(chat): host chip on the home composer — shared ComposerHostChip module
Extract the Host chip (popover + status dots + connect dialog) from
chat-view into src/components/composer-host-chip.tsx, self-contained
(owns its lazy /api/hosts fetch and the connect flow; the parent only
supplies value + onPick). Its styling moves to composer-host-chip.css
so both surfaces carry it without leaning on cave-chat.css.

The home composer renders the chip in its chat-only controls; a pick
rides initialControls.runtimeHost (InitialCommandControls extended)
into the opened chat, where it rides the FIRST send explicitly (the
state seed lands too late for that closure) and seeds the chat's own
Host chip. Local pick on home = auto (no field sent).

Pins: chat-view-polish repointed at the shared module + new home-
composer parity pins; home-composer's onStartChat pin updated.
