+++ src/components/home-composer.test.ts
@@ -63,8 +63,8 @@ assert.match(
 
 assert.match(
   source,
-  /onStartChat\(prompt, selectedFamiliarId, selectedProject\?\.root \?\? null, \{\s*initialControls: \{ thinkingEffort, responseSpeed \},[\s\S]*?\}\)/,
-  "HomeComposer should hand the selected project root and initial command controls to chat start",
+  /onStartChat\(prompt, selectedFamiliarId, selectedProject\?\.root \?\? null, \{\s*initialControls: \{ thinkingEffort, responseSpeed, \.\.\.\(runtimeHost \? \{ runtimeHost \} : \{\}\) \},[\s\S]*?\}\)/,
+  "HomeComposer should hand the selected project root, initial command controls, and any host pick to chat start",
 );
 
 assert.match(
